refactor (composite view): move utils into scripts dir
