--- utils/__init__.py
@@ -1,1 +1,0 @@
-from utils.helper_methods import absolute_from_rel_file_path
